Move UserPermission to domain, optimize imports in UserRole
